+++ lib/base/game_component.dart
@@ -22,10 +22,13 @@ abstract class GameComponent extends PositionComponent
   /// Param checks if this component is visible on the screen
   bool _isVisibleInScreen = false;
 
-  bool get isVisibleInScreen => visible ? _isVisibleInScreen : false;
+  bool get isVisible => _visible ? _isVisibleInScreen : false;
+  set isVisible(bool visible) {
+    _visible = visible;
+  }
 
   /// Param used to enable or disable the render.
-  bool visible = true;
+  bool _visible = true;
 
   bool enabledCheckIsVisible = true;
 
@@ -92,16 +95,16 @@ abstract class GameComponent extends PositionComponent
   }
 
   void _onSetIfVisible() {
-    if (!visible) return;
+    if (!_visible) return;
     bool nowIsVisible = isVisibleInCamera();
     if (isHud) {
       nowIsVisible = true;
       enabledCheckIsVisible = false;
     }
-    if (nowIsVisible && !isVisibleInScreen) {
+    if (nowIsVisible && !isVisible) {
       (gameRef as BonfireGame).addVisible(this);
     }
-    if (!nowIsVisible && isVisibleInScreen) {
+    if (!nowIsVisible && isVisible) {
       (gameRef as BonfireGame).removeVisible(this);
     }
     _isVisibleInScreen = nowIsVisible;