+++ lib/util/extensions/enemy/rotation_enemy_extensions.dart
@@ -18,7 +18,7 @@ extension RotationEnemyExtensions on RotationEnemy {
     bool runOnlyVisibleInScreen = true,
   }) {
     if (isDead) return;
-    if (runOnlyVisibleInScreen && !isVisibleInScreen) return;
+    if (runOnlyVisibleInScreen && !isVisible) return;
 
     seePlayer(
       radiusVision: radiusVision,
@@ -56,7 +56,7 @@ extension RotationEnemyExtensions on RotationEnemy {
     bool runOnlyVisibleInScreen = true,
   }) {
     if (isDead) return;
-    if (runOnlyVisibleInScreen && !isVisibleInScreen) return;
+    if (runOnlyVisibleInScreen && !isVisible) return;
 
     seePlayer(
       radiusVision: radiusVision,